+++ src/kaocha/type/var.clj
@@ -17,10 +17,7 @@
   (println "Test ran without assertions. Did you forget an (is ...)?")
   (report/print-output m))
 
-(defn test-var [{test    :kaocha.var/test
-                 the-var :kaocha.var/var
-                 meta'   :kaocha.testable/meta
-                 :as     testable}]
+(defn test-var [test the-var]
   (binding [t/*testing-vars* (conj t/*testing-vars* the-var)]
     (t/do-report {:type :begin-test-var, :var the-var})
     (try
@@ -28,21 +25,24 @@
       (catch clojure.lang.ExceptionInfo e
         (when-not (:kaocha/fail-fast (ex-data e))
           (report/report-exception e)))
-      (catch Throwable e (report/report-exception e)))
-    (let [{::result/keys [pass error fail pending] :as result} (type/report-count)]
-      (when (= pass error fail pending 0)
-        (binding [testable/*fail-fast?* false
-                  testable/*test-location* {:file (:file meta') :line (:line meta')}]
-          (t/do-report {:type ::zero-assertions}))))
-    (t/do-report {:type :end-test-var, :var the-var})
-    (merge testable {:kaocha.result/count 1} (type/report-count))))
+      (catch Throwable e (report/report-exception e)))))
 
-(defmethod testable/-run :kaocha.type/var [{wrap :kaocha.testable/wrap
-                                            :as  testable} test-plan]
+(defmethod testable/-run :kaocha.type/var [{test    :kaocha.var/test
+                                            wrap    :kaocha.testable/wrap
+                                            the-var :kaocha.var/var
+                                            meta'   :kaocha.testable/meta
+                                            :as     testable} test-plan]
   (type/with-report-counters
-    (let [wrapped-test (fn [] (test-var testable))
+    (let [wrapped-test (fn [] (test-var test the-var))
           wrapped-test (reduce #(%2 %1) wrapped-test wrap)]
-      (wrapped-test))))
+      (wrapped-test)
+      (let [{::result/keys [pass error fail pending] :as result} (type/report-count)]
+        (when (= pass error fail pending 0)
+          (binding [testable/*fail-fast?* false
+                    testable/*test-location* {:file (:file meta') :line (:line meta')}]
+            (t/do-report {:type ::zero-assertions}))))
+      (t/do-report {:type :end-test-var, :var the-var})
+      (merge testable {:kaocha.result/count 1} (type/report-count)))))
 
 (s/def :kaocha.type/var (s/keys :req [:kaocha.testable/type
                                       :kaocha.testable/id