+++ CHANGELOG.md
@@ -4,12 +4,13 @@
 
 ## Fixed
 
+- Ensure reloading errors are printed in watch mode when the first test suite is disabled.
 - Using a try-catch (without rethrowing) in an `:each` fixture could swallow
   thrown exceptions, resulting in a test being treated as passing when it should
   have been reported as an error. Fixed by changing how `:each` fixtures wrap
   the test function in execution. (thanks
   [@NoahTheDuke](https://github.com/NoahTheDuke))
-* Fix crash on Windows when using `--watch` with the default Beholder watcher.
+- Fix crash on Windows when using `--watch` with the default Beholder watcher.
 
 ## Changed