+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Added
 
+- `kaocha.repl/config` accepts a `:profile` key when specifying extra options.
+
 ## Fixed
 
 - Configuration errors are reported as warnings, rather than causing the entire