+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - Configuration errors and circular dependencies are reported as warnings, 
     rather than causing the entirewatcher to crash.
+- Fix bug added in [#384](https://github.com/lambdaisland/kaocha/issues/384):
+    assertions in the tail position of an each fixture would return the result
+    of the assertion instead of the testable object with the merged report data.
+    (thanks [@NoahTheDuke](https://github.com/NoahTheDuke))
 
 ## Changed