+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 1.76.1230 (2023-01-25 / 2b9d9df)
 
 ## Added